% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn.R
\name{createOffspring}
\alias{createOffspring}
\title{Create new offspring}
\usage{
createOffspring(genome1, genome2, min_val, max_val, p_mut = 0.25,
  w_mut = 0.2, p_cross = 0.1)
}
\arguments{
\item{genome1}{Vector containing the gene values for individual 1.}

\item{genome2}{Vector containing the gene values for individual 2.}

\item{min_val}{A vector of the minimum acceptable values for each gene.}

\item{max_val}{A vector of the maximum acceptable values for each gene.}

\item{p_mut}{Probability of a gene mutating (default: 0.25).}

\item{w_mut}{Maximum variation allowed through mutation (default: 0.25).}

\item{p_cross}{Probability of crossover (default: 0.1).}
}
\value{
A vector containing the gene values of the new individual.
}
\description{
Generate a new offspring from the genome of two parents.
}
\examples{
#TODO
}
\seealso{
\code{\link{}}
}

