% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneOptim.R
\name{geneOptim}
\alias{geneOptim}
\title{Quick & dirty genetic algorithm}
\usage{
geneOptim(fn, ..., N = 100, G = 500, min_val, max_val, p_mut = 0.25,
  w_mut = 0.2, p_cross = 0.1, n_cores = 1, .packages = NULL)
}
\arguments{
\item{fn}{Function to maximize. The ouput of the function should be a single
numeric value.}

\item{N}{Number of individual in the population (default: 100).}

\item{G}{Number of "generations". A generation actually corresponds to a
reproductive event between the best two individuals of a group of four
individuals randomly chosen in the population (default: 500).}

\item{min_val}{A vector of the minimum acceptable values for each gene.}

\item{max_val}{A vector of the maximum acceptable values for each gene.}

\item{p_mut}{Probability of a gene mutating (default: 0.25).}

\item{w_mut}{Maximum variation allowed through mutation (default: 0.25).}

\item{p_cross}{Probability of crossover (default: 0.1).}

\item{n_cores}{Number of available cores for parallel computation (default: 1).}
}
\value{
A matrix of the final population.
}
\description{
Maximization of a function using a quick & dirty algorithm.
}
\examples{
#TODO
}
\seealso{
\code{\link{}}
}

