% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn.R
\name{initPop}
\alias{initPop}
\title{Create initial population}
\usage{
initPop(N, min_val, max_val)
}
\arguments{
\item{N}{Number of individuals in the population.}

\item{min_val}{A vector of the minimum acceptable values for each gene.}

\item{max_val}{A vector of the maximum acceptable values for each gene.}
}
\value{
A N x length(min_val) matrix. Each line corresponds to an individual
 in the initial population. Each column correspond to the value of each gene,
 randomly chosen between min_val and max_val. The last column corresponds to
 the fitness of each individual (NA after initialization).
}
\description{
This function creates an initial population of randomly generated individuals.
}
\examples{
#TODO
}
\seealso{
\code{\link{}}
}

