% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn.R
\name{progress}
\alias{progress}
\title{Track optimization progress}
\usage{
progress(pop, k, min_fit, mean_fit)
}
\arguments{
\item{pop}{Matrix of the gene values of the current population. Each line
correspond to an individual in the initial population. Each column correspond
to the value of each gene. The last column corresponds to the fitness of each
individual.}

\item{k}{Number of genes.}

\item{minfit}{Vector of the minimum fitness for each generation.}

\item{meanfit}{Vector of the average fitness for each generation.}
}
\value{
Plot the minimum and average fitness for the last 100 generations.
}
\description{
Display current state and history of the genetic optimization.
}
\examples{
#TODO
}
\seealso{
\code{\link{}}
}

